Fix git repository discovery in npm version
When package.json is not in the git root, then npm version bumps didn't
make a commit.

Issue was with how npm version discovered if it's in a git repository: it
assumed that it's a git repository only if .git folder is at the same
level as the package.json.

In reality you cannot make such assumptions, as the .git directory is
implementation detail.

Proper way of discovering if you're in a git directory is invoking
`git rev-parse --git-dir`.